+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ Bugfixes
 ========
 
 * Fix regression causing plots for final model in AMD-tools to not show
+* Change behaviour of default in :code:`modeling.get_observations`, :code:`dv=None` now means using all DVIDs
 
 .. _2.0.0: